FBX loading - Objects now built in a proper graph. - Materials now loaded based on geometries parent model.

+++ Source/Tokamak.Readers/FBX/Builders/IFBXObject.cs
@@ -2,7 +2,7 @@
 {
     internal interface IFBXObject
     {
-        int ID { get; }
+        long ID { get; }
 
         string Name { get; }
